Add Markdown support.
Markdown support is optional. It is enabled by setting the `format`
property of the configuration object to "markdown."

ReSpec uses marked for parsing Markkdown.

Note that the content of SECTION elements, and P elements with a
class name of "note" or "issue" is also parsed.

The HTML created by the Markdown parser is turned into a nested
structure of SECTION elements, following the strucutre given by
the headings. For example, the following markup:

    Title
    -----

    ### Subtitle ###

    Here's some text.

    ### Another subtitle ###

    More text.

will be transformed into:

    <section>
      <h2>Title</h2>
      <section>
        <h3>Subtitle</h3>
        <p>Here's some text.</p>
      </section>
      <section>
        <h3>Another subtitle</h3>
        <p>More text.</p>
      </section>
    </section>