Allow issues to be marked atrisk
This represent a feature at risk. I also introduced a new configuration setting "atRiskBase" which allows similarly to "issueBase" to set the base URL to prepend to the issue number. Features at risk are often managed separately.

It would be great if this could released soon since we need it for the JSON-LD specification.

Thanks