Ensure that URL fragment identifiers work as intended
Currently, in most cases, URL fragment identifiers don't work as intended as the target doesn't exist yet. This module re-sets window.location.hash to jump to the right position in the document. It does this only if the document hasn't been scrolled down yet to ensure that page refreshes maintain the scroll position.