Revert to polling in respec2html.js
Commit 38e61f1a613278b2e2c6c823058dc00408ace149 moved respec2html.js
to waiting for the end-all message on respecEvents instead of polling
document.respecDone. However, the subscription is done in the
page.open callback, when ReSpec can already have finished, in which
case the tool will simply time out.

It's not clear if there is a page.on* callback which is guaranteed to
come after respecEvents.sub is available but before end-all is
published. None was found.

Also chmod -x the script, as it lacks the #! to be executable.