Addressed role of heading on h* elements
As per issue #370, the generation of role="heading"
on h* elements is not actually permitted by the HTML5
Recommendation and causes problems with some assistive
technology implementations.  Changed the strategy so
roles of heading are only placed on headings within notes and
issues.  In order to preserve the semantic data that explicit
roles can generate via RDFa, also added RDFa markup that maps
to the implied RDFa that would have been derived from the role
values were they still present.