Fixed header pollution from permalinks
hN elements were having permalink data included within them.  This made
them less accessible for people with screen readers.

In addition, the permalink label / title was such that it was difficult
to use assistive technology and keyboard shortcuts to navigate the links
to find a specific section.

The changes in this branch address the issues raised in #425