Implemented aliases for definitions
Added handling for aliases as per Bikeshed.  Specifically,
`<dfn title="foo|bar">my term</dfn>` is able to be referenced with
things like `<a>foo</a>` or `<a title="bar">my term</a>` or
`<a>bar</a>`.

This also ensures that the generated output is compatible with the
Bikeshed contract for external processors.  There is more work to do to
get that complete, but this is a good first step.

This is in general in reference to issue #344