Makes getDfnTitles work when called again
Some add-on code (used by PFWG and WPIG currently) call this utility
method on dfn elements again to get the title.  If there are aliases
for a dfn, these are moved into data-lt during the definition
processing phase, so a subsequent call will return the wrong list in
that case.  This fixes that (and improves the tests so we know it
works).