Improved Member-SUBM and Team-SUBM spec status support
When the spec status is "Member-SUBM" or "Team-SUBM", ReSpec expected a
"Previous version" link that does not mean anything in this context, did not
set the proper companion logo next to the W3C logo and did not include the
right Status of This Document boilerplate.

Please note that the links in the SoTD boilerplate for Member submissions still
contain "@@@". That is normal since these links are set by W3C Team right before
the submissions is accepted, and cannot be computed a priori.

ReSpec still conserves the default copyright in that case, but users simply
have to use "overrideCopyright" to override that setting.

This commit contains a few regression tests.