Changed logic to prefer data-lt for aliases
As per requests in issue #466 and #467, modified the logic for
finding link titles and aliases to prefer the use of the data-lt
attribute instead of 'lt' or 'title'.  'title' remains deprecated
so that we can remove support for it in some future version and
leave it available for A11Y uses.

If this is ultimately merged in, will also update the documentation
to reflect the preference for data-lt.