Added support for overriding the publication status
At the request of @michael-n-cooper I added the
respec option overrideStatus.  If this option is true,
then the paragraph that starts "This document was published by..."
will be omitted from the generated output of the SotD.  Apparently
some working groups object to some of the wording and want to provide
their own.