Fix ID of WebIDL enumeration value definitions in contiguous mode
@dontcallmedom's recent commits were not enough to fix the definition of
enumeration items. In practice, the template generated references using an
"idl-def-" prefix, preserving the case and using a dotted notation, while
definitions were made with a "dom-" prefix, lower case characters, and
dash-separated parts.

This commit ensures that the references produced in the WebIDL definition
correctly use the "dom-" prefix. It includes a regression test.