Added support for 'webspec' stylesheet.
Other support for webspec is already in w3c/headers.js. While the
webplatform.org project never really took off, there are certainly
documents out there that rely upon ReSpec and have this specStatus so we
need to leave support in for those.  We *could* issue a deprecation
warning, but it didn't feel that important to me.