Expand handlebars path arguments for build to run on Windows
As opposed to Unix systems, Windows does not automatically expand wildchars
passed as arguments in a command-line, and Handlebars does not handle that
expansion either.

This commit adds an expansion tool that wraps the "glob" library and passes
the result to handlebars. The build now works on Windows and the update should
not have any impact on Unix systems.