Make doc writer work with older versions of ReSpec
Version 20.10 introduced the `core/exporter` module, which is now used in
RespecDocWriter. Problem is the spec on which RespecDocWriter runs may
still reference an older version of ReSpec that does not have that module.

The writer checks the version of ReSpec and refuses to run on very old versions
(lower than version 18). It still accepts to run on versions between 18 and
20.10 (which is a good thing :)). The old `ui/save-html` module needs to be
used in such cases, which is what this update does.