Add a period and space to unlinked references
When a reference has a link, the stringifyReference function wraps the
title in the link and adds a ". " to the end of it to separate it from
the rest of the reference text. But if it doesn't have a link, neither
the link nor the ". " are added, causing the title to run into the next
portion of the output. This commit adds the separator for this case.
There may be more elegant ways to handle but this fix seems
straightforward.