Properly handle optional arguments in data-lt attributes
This update fixes the generation of the `data-lt` attribute when an operation
only has optional arguments. The code added a `, ` prefix before the list of
optional arguments, regardless of whether the operation had mandatory arguments
as well, leading to names such as `methodName(, options)` instead of
`methodName(options)` (see issue #3200).