Fixed: Ignore any elements that have SVG as a parent
This changes the behaviour of the issues and notes selector to ignore any elements
that have an SVG element as an ancestor. It does this by filtering the selected elements
and looking to see if the `.closest` method returns null; if it is null, it means that the element
is not within an SVG block; if it is not null, it is filtered out of the list of elements for further processing.

A simple unit test is included that checks to see if there are any <div> elements that are within
an <svg> element (<div> is not a valid child of <svg>). If the selector is null it means the test passed.

Fixes #4238.