Fixed: filter out non-HTMLElement instances
Changes the filter statement to filter out non-HTMLElement nodes instead
of looking for node ancestry.

Also fixes an issue with the linter and the wrong type arguments. Since `handleIssues` is
only used once (as far as I can tell) I think this is minimal. The type was changed since
we could not filter a `NodeList`, so it was changed to an array of `HTMLElement` objects.