feat(linter-rules/informative-dfn): warn on refs to informative dfns (#4329)
* Move warning on references to informative definitions to a separate linter

There are situations where making a normative reference to a definition in an informative section is either OK or unavoidable (e.g. because updating the referenced spec is unlikely to be possible). Moving the relevant check in a linter make it easier to silence either individual or all instances of the warning.