When normalising, fix `<` and `>` too
A fix for #4430 : when the normalized text is rendered as html e.g. in `dfn-index.js` line 298 within `renderExternalTermEntry` using `html: text` this prevents creation of what the UA thinks are HTML elements from terms defined with angle brackets at the start and end.

An alternative fix is to replace `html: text` with `text: text` but I think that would break cases where code blocks are deliberately put around the term in the index.

This fix seems to pass all the tests.