Preserve the filename in the URL of an external term
See #4787 for context. The URL of the spec may end with a filename but the code
assumed that it would always be a path, and thus that resolving `./` against
the URL of the spec would be a no-op. The code now makes sure resolve the URL
of the spec against an empty string when the xref search does not return a more
specific path.