Support dark mode toggle for SVGs and highlight.js
Syntax highlighting already works with automatic preferences, but has
not played well with neither default (no dark mode) settings, nor worked
with the manual scheme toggle. (See also #4871.)

This fixes that using CSS (so that default behavior works without JS).

SVG behavior is by default based on user-agent preferences, and has
therefore also not played well with the manual toggle. With this change,
that is now handled by setting `color-scheme: light` or `dark` on the
body element, based on the sidebar-provided manual scheme toggle.

If no meta element indicating dark color-scheme support is present,
light mode is set on the body. This instructs SVGs to render in light
mode when embedded in documents with no dark mode support, even when the
user agent indicates dark mode. (The `color-scheme` CSS property can of
course still be set to other values for various sections in a document,
if so desired.)

A darkmode example document is also added, containing syntax-highlighted
code and embedding an SVG document supporting light/dark color-scheme
preferences.