Report all `console.error` arguments
While attempting to debug an issue with the `respec-mermaid` plugin,
we discovered that the plugin logs multiple messages in the same
`console.error` call:

```
console.error('respec-mermaid error: Failed to generate figure.',
  e, mermaidSource);
```

Unfortunately only the first message would appear in the console
logs when running Respec. That's because it would get the `[0]`
index of the args array and would report that as text.

Instead of only reporting the first argument, instead loop over all
argument text parts and report them individually on the console.
We can then use this to figure out why our mermaid figures are
not correctly rendering on CI (they are fine on local).